+++ src/csv_value.rs
@@ -85,7 +85,9 @@ pub fn json_to_csv(json: &[u8]) -> Result<Vec<u8>> {
                         } else {
                             buffer.push(',');
                         }
-                        buffer.push_str(&value.to_string());
+                        if !value.is_null() {
+                            buffer.push_str(&value.to_string());
+                        }
                     } else {
                         bail!("Missing value for key: in {key:?}")
                     }